reconstruct k_fieldRuntime state1 interpolation pair

+++ src/Event/Fcl/fclShopMisc.c
@@ -16087,7 +16087,7 @@ u8 FUN_00403410(short param_1)
   int iVar3;
   int iVar4;
   long lVar5;
-  long unaff_s1;
+  int unaff_s1;
   long unaff_s2;
 
   iVar3 = FUN_004037e0(param_1);
